platformen hinzugefügt

+++ gdx-pneumatic-man/src/main/java/de/hochschuletrier/gdw/ws1415/game/utils/MapLoader.java
@@ -99,7 +99,7 @@ public class MapLoader
                     if(obj.getProperty("Name", "").equalsIgnoreCase("Platform")){
                         PlatformMode mode = PlatformMode.valueOf(obj.getProperty("Mode", PlatformMode.ALWAYS.name()).toUpperCase());
                         Direction dir = Direction.valueOf(obj.getProperty("Direction", Direction.UP.name()).toUpperCase()); // "Direction"
-                        int distance = obj.getIntProperty("Distance", 0);
+                        int distance = (int)obj.getFloatProperty("Distance", 0);
                         int hitpoints = obj.getIntProperty("Hitpoints", 0);
                         float speed = obj.getFloatProperty("Speed", 0);
                         if(hitpoints == 0)
